chore: level up haiku-formats typescriptification.
