+++ packages/haiku-formats/.gitignore
@@ -1,7 +1,6 @@
 node_modules
 .DS_Store
 *.log
-~test
 lib/
 .nyc_output/
 coverage/
