+++ packages/haiku-formats/package.json
@@ -7,18 +7,19 @@
     "lib": "lib"
   },
   "scripts": {
-    "lint": "tslint -p tsconfig.json -c tslint.json 'src/**/*.ts' -t stylish",
+    "lint": "yarn lint-src && yarn lint-test",
+    "lint-src": "tslint -p tsconfig.json -c tslint.json 'src/**/*.ts' -t stylish",
     "lint-test": "tslint -p tsconfig.test.json -c tslint.json 'test/**/*.ts' -t stylish",
-    "lint-report": "yarn lint -t checkstyle -o checkstyle-result.xml",
-    "fix": "yarn lint --fix",
-    "test": "tsc && tsc -p tsconfig.test.json && cross-env NODE_ENV=test tape \"~test/**/*.test.js\" | tap-spec && rm -rf ~test",
-    "test-report": "tsc -p tsconfig.test.json && cross-env NODE_ENV=test tape \"~test/**/*.test.js\" > test-result.tap",
+    "lint-report": "yarn lint-src -t checkstyle -o checkstyle-result.xml",
+    "fix": "yarn lint-src --fix && yarn lint-test --fix",
+    "ts": "ts-node -r tsconfig-paths/register -P tsconfig.test.json",
+    "test": "yarn ts ./node_modules/.bin/tape 'test/**/*.test.ts' | tap-spec",
+    "test-report": "yarn ts ./node_modules/.bin/tape 'test/**/*.test.ts' > test-result.tap",
     "coverage": "nyc yarn test",
-    "html-coverage": "nyc --reporter=html yarn test",
     "compile": "tsc",
     "develop": "tsc --watch",
-    "regenerate-goldens": "tsc && node ./scripts/regenerate-goldens.js",
-    "upgrade-goldens": "tsc && node ./scripts/upgrade-goldens.js",
+    "regenerate-goldens": "yarn ts ./scripts/regenerate-goldens.ts",
+    "upgrade-goldens": "yarn ts ./scripts/upgrade-goldens.ts",
     "demos": "node ./scripts/demos.js"
   },
   "authors": [
@@ -48,6 +49,8 @@
     "nyc": "^11.3.0",
     "tap-spec": "^4.1.1",
     "tape": "^4.8.0",
+    "ts-node": "^6.0.3",
+    "tsconfig-paths": "^3.3.2",
     "tslint": "^5.7.0",
     "tslint-config-haiku": "^1.0.0",
     "typescript": "^2.5.2",
