+++ packages/haiku-formats/test/exporters/curves.test.ts
@@ -1,9 +1,6 @@
-import * as tape from 'tape';
+import tape = require('tape');
 
-import {
-  decomposeCurveBetweenKeyframes,
-  splitBezierForTimelinePropertyAtKeyframe,
-} from '../../lib/exporters/curves';
+import {decomposeCurveBetweenKeyframes, splitBezierForTimelinePropertyAtKeyframe} from '@/exporters/curves';
 import {Curve} from '@haiku/core/lib/api/Curve';
 
 tape('bezier re-interpolation', (test: tape.Test) => {
