+++ packages/haiku-formats/tsconfig.test.json
@@ -1,12 +1,9 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "~test",
-    "rootDir": "test",
-    "declarationDir": "~test"
+    "noImplicitAny": false
   },
   "include": [
-    "test/**/*.ts",
-    "test/**/*.tsx"
+    "test/**/*.ts"
   ]
 }
